Store Staffing Plan by Process Path (#567)
- Stores process path information for process_distribution and headcount_productivity
- Changes process_distribution quantity type from long to double

+++ src/main/java/com/mercadolibre/planning/model/api/web/controller/forecast/request/ProcessingDistributionDataRequest.java
@@ -12,5 +12,5 @@ public class ProcessingDistributionDataRequest {
     @NotNull
     private ZonedDateTime date;
 
-    private long quantity;
+    private double quantity;
 }
